+++ src/Core/Credentials/Credentials.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Core\Credentials;
 
+use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
+use Bitrix24\SDK\Application\Requests\Placement\PlacementRequest;
+
 /**
  * Class Credentials
  *
@@ -23,6 +26,8 @@ class Credentials
      * @param AccessToken|null        $accessToken
      * @param ApplicationProfile|null $applicationProfile
      * @param string|null             $domainUrl
+     *
+     * @throws \Bitrix24\SDK\Core\Exceptions\InvalidArgumentException
      */
     public function __construct(
         ?WebhookUrl $webhookUrl,
@@ -33,6 +38,16 @@ class Credentials
         $this->webhookUrl = $webhookUrl;
         $this->accessToken = $accessToken;
         $this->applicationProfile = $applicationProfile;
+
+        if ($domainUrl !== null) {
+            $parseResult = parse_url($domainUrl);
+            if (!array_key_exists('scheme', $parseResult)) {
+                $domainUrl = 'https://' . $domainUrl;
+            }
+        }
+        if (($domainUrl !== null) && filter_var($domainUrl, FILTER_VALIDATE_URL) === false) {
+            throw new InvalidArgumentException(sprintf('domain URL %s is invalid', $domainUrl));
+        }
         $this->domainUrl = $domainUrl;
         if ($this->accessToken === null && $this->webhookUrl === null) {
             throw new \LogicException('you must set on of auth type: webhook or OAuth 2.0');
@@ -65,13 +80,11 @@ class Credentials
     {
         if ($this->getWebhookUrl() !== null) {
             $arUrl = parse_url($this->getWebhookUrl()->getUrl());
-
-            $url = sprintf('%s://%s', $arUrl['scheme'], $arUrl['host']);
         } else {
-            $url = $this->domainUrl;
+            $arUrl = parse_url($this->domainUrl);
         }
 
-        return $url;
+        return sprintf('%s://%s', $arUrl['scheme'], $arUrl['host']);
     }
 
     /**
@@ -94,6 +107,7 @@ class Credentials
      * @param WebhookUrl $webhookUrl
      *
      * @return self
+     * @throws \Bitrix24\SDK\Core\Exceptions\InvalidArgumentException
      */
     public static function createForWebHook(WebhookUrl $webhookUrl): self
     {
@@ -111,6 +125,7 @@ class Credentials
      * @param string             $domainUrl
      *
      * @return self
+     * @throws \Bitrix24\SDK\Core\Exceptions\InvalidArgumentException
      */
     public static function createForOAuth(AccessToken $accessToken, ApplicationProfile $applicationProfile, string $domainUrl): self
     {
@@ -119,6 +134,22 @@ class Credentials
             $accessToken,
             $applicationProfile,
             $domainUrl
+        );
+    }
+
+    /**
+     * @param \Bitrix24\SDK\Application\Requests\Placement\PlacementRequest $placementRequest
+     * @param \Bitrix24\SDK\Core\Credentials\ApplicationProfile             $applicationProfile
+     *
+     * @return self
+     * @throws \Bitrix24\SDK\Core\Exceptions\InvalidArgumentException
+     */
+    public static function initFromPlacementRequest(PlacementRequest $placementRequest, ApplicationProfile $applicationProfile): self
+    {
+        return self::createForOAuth(
+            $placementRequest->getAccessToken(),
+            $applicationProfile,
+            $placementRequest->getDomainUrl()
         );
     }
 }