+++ src/Services/Placement/Result/PlacementLocationItemResult.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Placement\Result;
+
+use Bitrix24\SDK\Core\Result\AbstractItem;
+
+/**
+ * @property-read string $placement
+ * @property-read string $handler
+ * @property-read string $title
+ * @property-read string $description
+ * @property-read array  $options
+ * @property-read array  $langAll
+ */
+class PlacementLocationItemResult extends AbstractItem
+{
+}